+++ package.json
@@ -28,11 +28,13 @@
     "body-parser": "1.19.0",
     "cors": "2.8.5",
     "express": "4.17.1",
+    "istanbul-instrumenter-loader": "3.0.1",
     "jasmine-core": "3.4.0",
     "js-polyfills": "0.1.42",
     "karma": "4.2.0",
     "karma-browserstack-launcher": "1.5.1",
     "karma-chrome-launcher": "3.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.5",
     "karma-jasmine": "2.0.1",
     "karma-spec-reporter": "0.0.32",
     "karma-typescript-preprocessor2": "1.2.1",
