+++ test/unit/karma.base.conf.js
@@ -1,8 +1,10 @@
+const path = require('path')
 const webpackConfig = require('../../webpack.config')(null, { mode: 'development' })
 
 module.exports = {
   basePath: '../..',
-  files: ['src/**/*.spec.ts'],
+  files: ['src/**/*.ts'],
+  exclude: ['src/**/*.d.ts'],
   frameworks: ['jasmine'],
   client: {
     jasmine: {
@@ -10,19 +12,23 @@ module.exports = {
     },
   },
   preprocessors: {
-    'src/**/*.spec.ts': ['webpack'],
+    'src/**/*.ts': ['webpack'],
   },
-  reporters: ['spec'],
+  reporters: ['coverage-istanbul', 'spec'],
   specReporter: {
     suppressErrorSummary: true,
     suppressPassed: true,
     suppressSkipped: true,
   },
+  coverageIstanbulReporter: {
+    reports: ['html', 'text-summary'],
+    dir: path.join(__dirname, '../../coverage'),
+  },
   singleRun: true,
   webpack: {
     mode: 'development',
     stats: 'minimal',
-    module: webpackConfig.module,
+    module: withIstanbulRule(webpackConfig.module),
     plugins: webpackConfig.plugins,
     resolve: webpackConfig.resolve,
   },
@@ -30,4 +36,19 @@ module.exports = {
     stats: 'errors-only',
     logLevel: 'warn',
   },
+}
+
+function withIstanbulRule(module) {
+  module.rules.push({
+    test: /^.*\.ts$/,
+    exclude: [/.*\.spec\.ts$/, /.*\.d\.ts$/, /.*capturedExceptions\.ts$/, /.*specHelper\.ts$/],
+    enforce: 'post',
+    use: {
+      loader: 'istanbul-instrumenter-loader',
+      options: {
+        esModules: true,
+      },
+    },
+  })
+  return module
 }
